move call span creation to client_tracing.go
